Rename manual_submission_script module to manual_operations
Refs #271
