+++ scripts/manual_operations/README.md
