+++ scripts/manual_operations/requirments.txt
